refactor(code-gen): 共享插件逻辑 + 框架适配器模式
