+++ packages/code-generator/src/plugins/component/shared/handlers/actions/default.ts
@@ -1,7 +1,5 @@
-// src/code-generator/plugins/component/react/handlers/action/default.ts
-
 /**
- * @file 默认 Action 处理器
+ * @file 默认 Action 处理器（框架无关）
  * @description 当未找到特定 actionType 的处理器时，回退到此处理器并输出调试日志。
  */
 
@@ -13,7 +11,6 @@ import type { IActionHandler } from "./type";
  * @returns 一行 `console.log` 调用代码字符串
  */
 export const defaultActionHandler: IActionHandler = (action) => {
-  // 默认的 console.log
   return `console.log('Action triggered: ${
     action.actionType
   }', ${JSON.stringify(action.config)});`;
