+++ packages/code-generator/src/plugins/component/shared/handlers/actions/goToLink.ts
@@ -1,5 +1,5 @@
 /**
- * @file 跳转链接 Action 处理器
+ * @file 跳转链接 Action 处理器（框架无关）
  * @description 将 schema 中配置的 URL 转换为 window.open 调用，用于在新标签页中打开链接。
  */
 
@@ -10,7 +10,7 @@ import type { IActionHandler } from "./type";
  * @param action - 当前 IRAction，其中 `config.url` 为要打开的地址
  * @returns `window.open` 调用代码字符串
  */
-export const goToLinkHander: IActionHandler = (action) => {
+export const goToLinkHandler: IActionHandler = (action) => {
   const url = JSON.stringify(action.config.url);
-  return `window.open(${url}, "_blank)`;
+  return `window.open(${url}, "_blank");`;
 };
