--- packages/code-generator/src/plugins/component/vue/handlers/actions/callMethod.ts
@@ -1,15 +1,0 @@
-/**
- * @file Vue 调用页面方法 Action 处理器
- */
-
-import type { IVueActionHandler } from "./type";
-
-export const callMethodHandler: IVueActionHandler = (action) => {
-  let callName = action.config.methodName; // "this.methods.open_123"
-
-  if (callName.startsWith("this.methods.")) {
-    callName = callName.substring("this.methods.".length);
-  }
-
-  return `${callName}();`;
-};
