--- packages/code-generator/src/plugins/component/vue/handlers/actions/goToLink.ts
@@ -1,10 +1,0 @@
-/**
- * @file Vue 跳转链接 Action 处理器
- */
-
-import type { IVueActionHandler } from "./type";
-
-export const goToLinkHandler: IVueActionHandler = (action) => {
-  const url = JSON.stringify(action.config.url);
-  return `window.open(${url}, "_blank");`;
-};
