+++ packages/code-generator/src/plugins/project/index.ts
@@ -1,31 +1,16 @@
 // src/code-generator/plugins/project/index.ts
 
 /**
- * @file 项目级出码插件入口
- * @description 汇总并导出所有项目级插件（样式、路由、入口文件、配置等），由解决方案统一执行。
- *
- * 注意：5 个静态模板文件（tsconfig、vite-config、gitignore、index-html、vite-env）
- * 已迁移到 ReactViteTemplate，不再作为独立插件。
+ * @file 项目级插件兼容入口
+ * @description 历史上该目录默认承载 React 插件；现已按框架拆分为 react/ 与 vue/。
+ * 当前文件仅作为兼容桥接，避免旧引用路径失效。
  */
 
-import globalStylePlugin from "./global-style";
-import routerPlugin from "./router";
-import entryPlugin from "./entry";
-import packageJsonPlugin from "./package-json";
-import componentsPlugin from "./components";
-
-export const projectPlugins = [
-  globalStylePlugin,
-  routerPlugin,
-  entryPlugin,
-  packageJsonPlugin,
-  componentsPlugin,
-];
-
 export {
+  projectPlugins,
   globalStylePlugin,
   routerPlugin,
   entryPlugin,
   packageJsonPlugin,
   componentsPlugin,
-};
+} from "./react";
