+++ packages/code-generator/src/plugins/project/react/tsconfig.ts
@@ -1,4 +1,4 @@
-// src/code-generator/plugins/project/tsconfig.ts
+// src/code-generator/plugins/project/react/tsconfig.ts
 
 import type { IProjectPlugin } from "@lowcode/schema";
 import type { ProjectBuilder } from "@lowcode/schema";
