+++ packages/code-generator/src/templates/react-vite-template.ts
@@ -1,15 +1,11 @@
 /**
  * @file React + Vite 项目模板
  * @description 实现 IProjectTemplate 接口，提供不依赖 IR 的静态脚手架文件。
- *              将原先 5 个静态 project plugin（tsconfig/vite-config/gitignore/index-html/vite-env）的内容
- *              整合到此模板中。
  */
 
 import type { IProjectTemplate, IGeneratedFile } from "@lowcode/schema";
+import { getTsconfigNodeJson, getGitignore } from "./shared";
 
-/**
- * React + Vite 项目模板
- */
 export const reactViteTemplate: IProjectTemplate = {
   name: "react-vite",
 
@@ -49,26 +45,8 @@ export const reactViteTemplate: IProjectTemplate = {
         fileType: "json",
       },
 
-      // --- tsconfig.node.json ---
-      {
-        fileName: "tsconfig.node.json",
-        filePath: "tsconfig.node.json",
-        content: JSON.stringify(
-          {
-            compilerOptions: {
-              composite: true,
-              skipLibCheck: true,
-              module: "ESNext",
-              moduleResolution: "bundler",
-              allowSyntheticDefaultImports: true,
-            },
-            include: ["vite.config.ts"],
-          },
-          null,
-          2,
-        ),
-        fileType: "json",
-      },
+      // --- tsconfig.node.json (共享) ---
+      getTsconfigNodeJson(),
 
       // --- vite.config.ts ---
       {
@@ -116,93 +94,8 @@ export default defineConfig({
         fileType: "html",
       },
 
-      // --- .gitignore ---
-      {
-        fileName: ".gitignore",
-        filePath: ".gitignore",
-        content: `# Logs
-logs
-*.log
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-pnpm-debug.log*
-lerna-debug.log*
-
-# Runtime data
-pids
-*.pid
-*.seed
-*.pid.lock
-
-# Directory for instrumented libs generated by jscoverage/JSCover
-lib-cov
-
-# Coverage directory used by tools like istanbul
-coverage
-*.lcov
-
-# nyc test coverage
-.nyc_output
-
-# Grunt intermediate storage (http://gruntjs.com/creating-plugins#storing-intermediate-files)
-.grunt
-
-# node-waf configuration
-.lock-wscript
-
-# Compiled binary addons (http://nodejs.org/api/addons.html)
-build/Release
-
-# Dependency directories
-node_modules/
-jspm_packages/
-
-# Snowpack dependency directory (https://snowpack.dev/)
-web_modules/
-
-# TypeScript cache
-*.tsbuildinfo
-
-# Optional npm cache directory
-.npm
-
-# Optional eslint cache
-.eslintcache
-
-# Microbundle cache
-.rpt2_cache/
-.rts2_cache_cjs/
-.rts2_cache_es/
-.rts2_cache_umd/
-
-# Optional REPL history
-.node_repl_history
-
-# Output of 'npm pack'
-*.tgz
-
-# Yarn Integrity file
-.yarn-integrity
-
-# dotenv environment variables file
-.env
-.env.development.local
-.env.test.local
-.env.production.local
-.env.local
-
-# Vite build output
-dist
-
-# Vite cache
-.vite
-
-# IDE directories
-.idea/
-.vscode/`,
-        fileType: "other",
-      },
+      // --- .gitignore (共享) ---
+      getGitignore(),
 
       // --- src/vite-env.d.ts ---
       {
